requirements.txt
# FoodScout AI — Python Dependencies

# Web framework
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1

# Async HTTP
httpx==0.27.0
aiohttp==3.9.5
requests==2.32.3

# Scraping
playwright==1.44.0
beautifulsoup4==4.12.3
lxml==5.2.2

# YouTube
google-api-python-client==2.130.0
youtube-transcript-api==0.6.2

# AI / LLM
openai==1.30.5

# Background tasks
celery==5.4.0
redis==5.0.4
flower==2.0.1

# Data processing
rapidfuzz==3.9.1
pandas==2.2.2

# Google Sheets export
gspread==6.1.2
google-auth==2.30.0
google-auth-oauthlib==1.2.0

# Config
python-dotenv==1.0.1
pydantic==2.7.4
pydantic-settings==2.3.0

# Utilities
tenacity==8.3.0
loguru==0.7.2
python-slugify==8.0.4
